Add linters configuration (#2)
Parent story: https://github.com/fenya123/bingin/issues/9

Here at fenya123 we don't settle for anything other than perfection. That is why our ACIH's backend code has to be uniform and consistent from start to finish. To achieve that, we are going to use linters: mypy, ruff, flake8, pylint.

In the world of coding speed is everything. The slower you are, the faster you and your work become obsolete. Following linters have overlapping checks and also vary in speed (fastest to slowest): ruff, flake8, pylint.

Naturally, the need for speed necessitates us to get rid of overlapping checks. We'll achieve that by configuring faster linters to perform as many checks as possible, while slower linters will only perform checks that others linters can't.

In the scope of this task we'll need to:

add and configure our linters of choice and their plugins,
add instructions to README.md on how to run them.