Add Docker caching to CI (#17)
Parent story: https://github.com/fenya123/bingin/issues/12

Now that we've added both linters and test to our CI it is time for optimization. At the moment, our CI pipeline builds its' containers from ground up on every push, which is time consuming. In order to save time we are going to add Docker caching to our CI. It will allow us to use cached layers instead of computing all of them (even unchanged ones) every time.

In the scope of this task we need to:

1. write an action that will setup buildx and cache its' cache
2. create dotenv files and write an action that will update buildx cache (workaround for [that](https://docs.docker.com/build/ci/github-actions/cache/#local-cache))
3. update our workflows and docker compose files accordingly