Add cleanup script for removed branch (#21)
Parent story: https://github.com/fenya123/bingin/issues/13

Removed branches leave behind workflow runs that are no longer needed.

In the scope of this task we will write a python script that deletes
workflow runs of a deleted branch. We are going to use a
[Personal access token](https://github.com/settings/tokens) in order to
log in to GitHub because authentication through login-password pair
doesn't work at the moment.