Improve cleanup workflow to remove its' previous runs (#29)
Parent story: https://github.com/fenya123/bingin/issues/13

The only thing our cleanup workflow lacks is self-cleanup. Every time our `cleanup` workflow is triggered it leaves behind a workflow run that accumulate with time. We only need one, so redundant workflow runs need to be eliminated.

In the scope of this task we are going to add a job to our `cleanup` Workflow that deletes previous runs of `cleanup` workflow.