Add linter configurations specific for FastAPI and Pydantic (#38)
Parent story: https://github.com/fenya123/bingin/issues/14

Since we are going to use FastAPI and Pydantic extensively during backend development, we need to prepare our linters for that. Because both of these libraries have some "features" which some linters cannot understand without additional plugins/configurations.

Configurations added in this ticket came from experience of working with FastAPI-Pydantic-linters setup. None of them we need at exact this moment, but they will be required after a while.

So in this ticket we need to add configuration which will "explain" some features of Pydantic and FastAPI to linters we use.