Add PostgreSQL as local container (#37)
Parent story: https://github.com/fenya123/bingin/issues/15

Next main development tool on the menu is database. In the scope of this task we need to add PostgreSQL to our project as a local container with attached volume.

Acceptance criteria:
1. a local environment directory with `docker-compose.yml` file are  created
2. `docker exec -it acih-local-dev-postgres-1 /bin/bash` launches `bash` inside container
3. `psql -U <username>` launches `Postgres` inside container
4. ```CREATE TABLE entity (id integer);  INSERT INTO entity VALUES (42);  SELECT * FROM entity;```
returns inserted values **both at the first container run AND after a restart**
5. previously created table cleaned via `DROP TABLE entity;`