Pytest integration (#47)
Parent story: https://github.com/fenya123/bingin/issues/15

To ensure long term development efficiency we need our application thoroughly tested. After configuring our framework and database to work together (#45) we can now get to testing. We want our tests to run on a clean database and to cleanup after they're done.

In the scope of this task we will:

1. Set up separate environment for testing with its' own container and volume.
2. Write simple tests for our existing routes.
3. Configure our testing setup in a way that every test is going to be isolated and independent. We will achieve that through using `pytest fixtures` in conjunction with `FastAPI's dependency overrides`.

NOTES:
Regarding implementation: we want our tests to 1) run on a clean database and 2) clean after themselves. We could achieve it by passing our session to tests and repeating the process of `Base.metadata.drop_all` and `...create_all`, then `session.add`ing and `session.commit`ing all the data we need for a given test. We could even encapsulate that process using fixtures to avoid repetitive code. That approach however has significant downsides:
1. `drop_all` method has its' caveats (such as static data or related columns interfering with dropping process)
2. recreating the whole database for each test will inevitably cause performance issues down the line.

A better option will be to provide our tests with a new session that will rollback all changes made with each test run. Considering that every test will require database (empty or otherwise), that session will have to be a `pytest fixture`. With this [approach](https://docs.sqlalchemy.org/en/20/orm/session_transaction.html#joining-a-session-into-an-external-transaction-such-as-for-test-suites) we will run database operations during tests inside a transaction that is going to get rolled back at the end of each test, erasing any data created during it. Code after `yield` is considered a "teardown" and will be executed after the test is done.
Our main session is injected into rout functions via `FastAPI's dependency`, so for our tests to work we need to [override](https://fastapi.tiangolo.com/advanced/testing-database/) it with our new session. And because our new session is inside a fixture as well as `TestClient` we unite them inside one.