Add configuration (#49)
Parent story: https://github.com/fenya123/bingin/issues/15

Some variable values (i.e. database URIs) in our application are hard-coded. It was fine so far, but hard-coding environment variables was of course a temporary measure. It's time to add configuration for our application.

In the scope of this task we will:

1. Create a configuration class via subclassing `Pydantic`'s `BaseSettings`
2. Configure this class in a way to recieve, parse and validate a `ENV_FILE` variable that is going to determine which dotenv file the configuration object is going to get its' values from
3. Change all hardcoded values to the values from config