Add deploy workflow (#75)
Parent story: https://github.com/fenya123/bingin/issues/17

We want to automate our deployment process. We will be using GitHub Actions for that purpose. We are going to add a `deploy.yml` workflow that will trigger on a certaing Git tag push and execute commands on VPS. In order to access VPS from Actions platform we will be using [Appleboy](https://github.com/appleboy/ssh-action) Action.

As a part of (#73) we already done:

1. Download Docker on VPS
2. Create SSH key on VPS and link it to GitHub account
3. CLone project's repo to VPS

In the scope of this task we will:

4. Add VPS's secrets to repository
5. Add `deploy.yml` workflow, which will
   - be triggered by pushing `DEPLOYED/QA` tag
   - run script on VPS via SSH which will update our app
7. Update README.md