Pin versions for external dependencies (#83)
Parent story: https://github.com/fenya123/bingin/issues/20

We need to pin versions for external dependencies to avoid version conflicts and other complications in the future (i.e. we've been working with postgres version 15 so far, but after version 16 came out volumes that were created with a previous version stopped working).

In the scope of this task we will:

1. Change our external dependenies (docker, minio) versions from latest to a fixed one.