Update FastAPI and Pydantic (#103)
Parent story: https://github.com/fenya123/bingin/issues/29

Back when we were deciding what version of FastAPI to add to our project we had to choose an older one due to some techincal difficulties. It was whole major patch behind, and because FastAPI relies heavily on Pydantic we also added an older version of it (v1 instead of v2). Now, standing on a brink of adding actual enpoint functionality we've decided to update to current versions of FastAPI and Pydantic as aforementioned difficulties were resolved.

In the scope of this task we will:
1. Update FastAPI and Pydantic versions in `pyproject.toml` and resolve dependencies
2. Migrate codebase accordingly