Add functionality for create_account endpoint (#105)
Parent story: https://github.com/fenya123/bingin/issues/19

Now we get to adding functionality to our endpoints. The first one to be written is create_account. With accordance to MVC design pattern, route fucntion's contents are to be moved into a separate module (controller), which in turn will evoke models, schemas and methods required for it's endpoint.

During this task three models are going to be added: Account, Profile and PasswordHash. Although our endpoint is supposed to create an account, our sign in\up\out system is designed in a way that no profile can exist without account and vice versa, same for password hash and account. Thus, endpoint will also implicitly create these entities and it is decided that Account will be the main entity that evokes creation of Profile and PasswordHash.

In the scope of this task we will:

1. Add Account, Profile and PasswordHash models in accordance with DDD
2. Add methods to models for creating them
3. Generate database migration
4. Update Pydantic schemas to validate incoming data in accordance with DDD
5. Write tests for our endpoint