Add functionality for upload_file endpoint (#107)
Parent story: https://github.com/fenya123/bingin/issues/28

In this task we start adding functionality to our `files` feature. Uploading and downloading images to and from our web application is one of it's most fundamental features.
There are certain notable descisions we had to make:
1. According to our DDD model, files on our platform might measure all the way up to 200 Mb. Considering this, it was decided to implement file upload functionality in a way that allows data to travel to and from our storage using as few steps as possible. Meaning, we want to omit opening/reading file data if we can. We want to come as close to operating on a raw bytes stream as our instruments allow us.
2. Instead of trying to catch and handle exceptions ourselves, we try to delegate error handling to Pydantic as much as possible (one of the reasons why we use it)

In the scope of this task we will:

1. Add `File` model
2. Generate migration
3. Add utilities (metrics, schemas etc)
4. Add endpoint functionality (encapsulated in a controller)
5. Add tests

P.S. We used to have a `minio` feature that we implemented in order to test MinIO. Now that we have tested endpoints that use MinIO, we no longer need that feature. Thus, we get rid of it in the scope of this task.