Add functionality for create_following endpoint (#117)
Parent story: https://github.com/fenya123/bingin/issues/24

With this task we start developing functionality for our Following endpoints. This feature's design compels us to set certain boundaries such as inabilty to follow self and follow someone twice. These checks/rules are to be placed both on a database as well as in the schemas/controller function.

In the scope of this task we will:

1. Go through usual development process (generating migrations, implementing controller functions, adding model methods etc)