Add functionality for create_post endpoint (#125)
Parent story: https://github.com/fenya123/bingin/issues/23

With this task we start developing functionality for our `Post` endpoints. During this task we stumbled upon a limitation of our chosen design. This endpoint's controller function has many of our project's facets intertwined to the point where we had to make a choice: either continue with an ugly controller or compromise code integrity (just a little bit). We've took the latter turn by coming up with a decent workaround.

In the scope of this task we will:

1. create `Post` model
2. generate database migration
3. add functionality (controller function, dependencies, model methods, router etc.)
4. add tests for new code