Add errors hierarchy (#141)
Parent story: https://github.com/fenya123/bingin/issues/33

Our error handling is still bare bones. As of now, we basically have a bunch of exceptions with hand-written descriptions sprinkled throughout the project. What we need is a unified system with basic exceptions and handlers that will handle their subclasses. That way, we can change any exception we need in one place instead of seeking it by hand in the codebase.

In the scope of this task we will:

1. Define base classes for exceptions
2. Write handlers for them
3. Add specific exceptions subclassed for base ones
4. Update the code/tests